@@ -1,20 +1,23 @@
 package model;
 
+import annotation.InjectEntityMangaer;
 import data.Repository;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
 import org.hibernate.cfg.Configuration;
-import org.junit.rules.MethodRule;
-import org.junit.runners.model.FrameworkMethod;
-import org.junit.runners.model.Statement;
+import org.junit.jupiter.api.extension.*;
+
 
 import javax.resource.cci.Record;
+import java.lang.reflect.Field;
+import java.util.Arrays;
+import java.util.stream.Collectors;
 
-public class SessionFactoryRule implements MethodRule {
+public class SessionFactoryExtension implements Extension, BeforeEachCallback, AfterEachCallback,BeforeAllCallback {
     private SessionFactory sessionFactory;
     private Transaction transaction;
     private org.hibernate.Session session;
-
+/*
     @Override
     public Statement apply(final Statement statement, FrameworkMethod method,
                            Object test) {
@@ -32,6 +35,47 @@ public class SessionFactoryRule implements MethodRule {
             }
         };
     }
+
+
+ */
+
+    @Override
+    public void beforeAll(ExtensionContext extensionContext) throws Exception {
+    }
+
+    private void injectEntityManagers(Object object) {
+        Class<?> aClass = object.getClass();
+        Arrays.stream(aClass.getDeclaredFields()).filter(lvField -> lvField.getAnnotationsByType(InjectEntityMangaer.class).length > 0)
+                .forEach(lvField -> {
+                    try {
+                        injectInterface(object, lvField);
+                    } catch (IllegalAccessException e) {
+                        e.printStackTrace();
+                    }
+                });
+    }
+
+    private void injectInterface(Object object, Field lvField) throws IllegalAccessException {
+        if (lvField.getAnnotationsByType(InjectEntityMangaer.class).length > 0) {
+            lvField.setAccessible(true);
+            Repository lvRepository = (Repository) lvField.get(object);
+            injectManager(lvRepository);
+        }
+    }
+
+    @Override
+    public void beforeEach(ExtensionContext extensionContext) throws Exception {
+        sessionFactory = createSessionFactory();
+        extensionContext.getTestInstance().ifPresent(this::injectEntityManagers);
+        createSession();
+        beginTransaction();
+    }
+
+    @Override
+    public void afterEach(ExtensionContext extensionContext) throws Exception {
+      //  shutdown();
+    }
+
     private void shutdown() {
         try {
             try {
@@ -77,7 +121,7 @@ public class SessionFactoryRule implements MethodRule {
         return session;
     }
 
-    public void injectManager(Repository repository){
+    private void injectManager(Repository repository){
         repository.setEntityManager(createSession().getEntityManagerFactory().createEntityManager());
     }
 }
