@@ -5,11 +5,10 @@ import lombok.extern.slf4j.Slf4j;
 
 import java.util.Collections;
 import java.util.UUID;
-import org.junit.Assert;
-import org.junit.Rule;
-import org.junit.Test;
 
-import static org.junit.Assert.*;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
 
 @Slf4j
 public class SessionTest {
@@ -20,16 +19,16 @@ public class SessionTest {
          String one = Session.generateRandom(Session.scale);
         String two = Session.generateRandom(Session.scale);
 
-        Assert.assertTrue(one.length() == Session.scale);
-        Assert.assertTrue(two.length() == Session.scale);
-        Assert.assertFalse(one.equals(two));
+        Assertions.assertTrue(one.length() == Session.scale);
+        Assertions.assertTrue(two.length() == Session.scale);
+        Assertions.assertFalse(one.equals(two));
 
         int as = 0;
         for(int i=0; i<one.length(); i++){
             if (one.charAt(i) == two.charAt(i)) as++;
         }
 
-        Assert.assertTrue(as<40);
+        Assertions.assertTrue(as<40);
 
         log.info(one);
         log.info(two);
@@ -47,7 +46,7 @@ public class SessionTest {
         two.setId(one.getId());
         two.setRecords(one.getRecords());
 
-        Assert.assertTrue(one.equals(two));
+        Assertions.assertTrue(one.equals(two));
     }
 
 }
