@@ -1,35 +1,30 @@
 package data;
 
+import annotation.InjectEntityMangaer;
+import model.Record;
 import model.*;
+import org.joda.time.LocalDateTime;
+import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-import org.hibernate.validator.cfg.defs.AssertTrueDef;
-import org.joda.time.LocalDate;
-import org.joda.time.LocalDateTime;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-
-import static junit.framework.TestCase.fail;
-
+@ExtendWith(SessionFactoryExtension.class)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class RecordRepositoryTest {
 
-    @Rule
-    public SessionFactoryRule sessionFactoryRule = new SessionFactoryRule();
-
+    @InjectEntityMangaer
     public RecordRepositoryImpl recordRepository;
+
+    @InjectEntityMangaer
     public SessionRepositoryImpl sessionRepository;
 
-    @Before
+    @BeforeAll
     public void init(){
         sessionRepository = new SessionRepositoryImpl();
         recordRepository = new RecordRepositoryImpl();
-        sessionFactoryRule.injectManager(recordRepository);
-        sessionFactoryRule.injectManager(sessionRepository);
     }
 
     @Test
@@ -43,8 +38,8 @@ public class RecordRepositoryTest {
         clone.setMediaType(record.getMediaType());
         clone.setLink(record.getLink());
         clone.setSession(record.getSession());
-        Assert.assertTrue(clone.toString().equals(record.toString()));
-        Assert.assertTrue(record.equals(clone));
+        Assertions.assertTrue(clone.toString().equals(record.toString()));
+        Assertions.assertTrue(record.equals(clone));
     }
 
     @Test
@@ -55,11 +50,11 @@ public class RecordRepositoryTest {
         Record save = recordRepository.save(record);
         recordRepository.getEntityManager().getTransaction().commit();
 
-        Assert.assertNotNull(save);
+        Assertions.assertNotNull(save);
 
         Record byId = recordRepository.getById(record.getId());
-        Assert.assertNotNull(save);
-        Assert.assertTrue(record.equals(save));
+        Assertions.assertNotNull(save);
+        Assertions.assertTrue(record.equals(save));
     }
 
     @Test
@@ -89,16 +84,16 @@ public class RecordRepositoryTest {
         recordRepository.getEntityManager().getTransaction().commit();
 
         List<Record> notDownloaded = recordRepository.getNotDownloaded();
-        Assert.assertNotNull(notDownloaded);
-        Assert.assertTrue(savedNotDownloaded.size() == notDownloaded.size() );
+        Assertions.assertNotNull(notDownloaded);
+        Assertions.assertTrue(savedNotDownloaded.size() == notDownloaded.size() );
 
         for (Record rec : savedNotDownloaded){
-            Assert.assertTrue(notDownloaded.contains(rec));
+            Assertions.assertTrue(notDownloaded.contains(rec));
         }
 
         for (int i =0; i<notDownloaded.size(); i++){
             if (i > 0){
-                Assert.assertTrue(notDownloaded.get(i-1).getDate().before(notDownloaded.get(i).getDate()));
+                Assertions.assertTrue(notDownloaded.get(i-1).getDate().before(notDownloaded.get(i).getDate()));
             }
         }
 
@@ -131,15 +126,15 @@ public class RecordRepositoryTest {
         recordRepository.getEntityManager().getTransaction().commit();
 
         List<Record> notDownloaded = recordRepository.getNotDownloaded(3);
-        Assert.assertNotNull(notDownloaded);
-        Assert.assertTrue(notDownloaded.size() == 3 );
+        Assertions.assertNotNull(notDownloaded);
+        Assertions.assertTrue(notDownloaded.size() == 3 );
 
         for (Record rec : notDownloaded){
-            Assert.assertTrue(savedNotDownloaded.contains(rec));
+            Assertions.assertTrue(savedNotDownloaded.contains(rec));
         }
 
-        Assert.assertTrue(notDownloaded.get(0).getDate().before(notDownloaded.get(1).getDate()));
-        Assert.assertTrue(notDownloaded.get(1).getDate().before(notDownloaded.get(2).getDate()));
+        Assertions.assertTrue(notDownloaded.get(0).getDate().before(notDownloaded.get(1).getDate()));
+        Assertions.assertTrue(notDownloaded.get(1).getDate().before(notDownloaded.get(2).getDate()));
 
 
     }
@@ -166,7 +161,7 @@ public class RecordRepositoryTest {
         recordRepository.getEntityManager().getTransaction().commit();
 
         Record byId = recordRepository.getById(object.getId());
-        Assert.assertNotNull(byId.getDate());
+        Assertions.assertNotNull(byId.getDate());
         System.out.println(byId.getDate().toString());
     }
 
@@ -182,16 +177,17 @@ public class RecordRepositoryTest {
 
         }
         List<Record> first = recordRepository.get5First();
-        Assert.assertTrue(first != null);
-        Assert.assertTrue(first.size() == 5);
+        Assertions.assertTrue(first != null);
+        Assertions.assertTrue(first.size() == 5);
         LocalDateTime localDateTime = LocalDateTime.fromDateFields(first.get(0).getDate());
         LocalDateTime readablePartial = LocalDateTime.fromDateFields(first.get(1).getDate());
         System.out.println(localDateTime.toString() + " " + readablePartial.toString());
-        Assert.assertTrue(localDateTime.isBefore(readablePartial));
+        Assertions.assertTrue(localDateTime.isBefore(readablePartial));
     }
-    @Test(expected = IllegalArgumentException.class)
+    @Test
     public void nullSave(){
-        sessionRepository.save(null);
+        Assertions.assertThrows(IllegalArgumentException.class, () -> sessionRepository.save(null));
+
     }
 
     @Test
@@ -206,7 +202,7 @@ public class RecordRepositoryTest {
         Record save = recordRepository.save(record);
         recordRepository.getEntityManager().getTransaction().commit();
 
-        Assert.assertTrue(record.getMediaType().equals(save.getMediaType()));
-        Assert.assertTrue(record.equals(save));
+        Assertions.assertTrue(record.getMediaType().equals(save.getMediaType()));
+        Assertions.assertTrue(record.equals(save));
     }
 }
